MCAL_BOT_Chatbot/requirements.txt
gradio==3.39.0
jose==1.0.0
passlib==1.7.4
PyPDF2==3.0.1
aspose-words==23.7.0
bs4==0.0.1
fastapi==0.100.1
requests==2.31.0
uvicorn==0.23.2
mysql-connector-python==8.1.0
Flask==2.3.2
PyJWT==2.8.0